llm-code-review-assistant/requirements.txt
httpx>=0.27.0
openai>=1.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pytest>=8.2.0
rich>=13.7.1
